Normal-inverse-Wishart distribution

Density and related functions for the normal-inverse-wishart distribution.

See Details.

See Details.

Usage

dnorminvwishart(mu, mu0, lambda, Sigma, S, nu, log = FALSE)

rnorminvwishart(n = 1, mu0, lambda, S, nu)

Arguments

mu

data or parameters vector of length \(k\) or matrix with \(k\) columns.

mu0

mean vector \(\mu_0\).

lambda

positive-only scalar.

Sigma

\(k \times k\) covariance matrix.

S

symmetric, positive-semidefinite \(k \times k\) scale matrix.

nu

scalar degrees of freedom \(\nu\).

log

logical; if TRUE, the logarithm of the density is returned.

n

number of random draws.

Value

dnorminvwishart gives the density, and other functions provide related computations.

See Details.

See Details.

Details

Implementation of dnorminvwishart. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of rnorminvwishart. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

See also

Examples

if (FALSE) { # \dontrun{
## see package vignettes for full examples involving dnorminvwishart
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving rnorminvwishart
} # }