Is ssm model

S3 method: apply is() to objects of class ssm_model.

These functions each provide a logical test of the class of an object.

Functions in Lucifer often assign a class to an output object. For example, after updating a model with the lucifer function, an output object is created of class demonoid. Likewise, after passing a model to the LaplaceApproximation function, an output object is created, and it is of class laplace. The class of these and other objects may be logically tested.

By assigning a class to an output object, the package is able to discern which other functions are appropriate for it. For example, after updating a model with lucifer, which creates an object of class demonoid, the user may desire to plot its output. Since it is assigned a class, the user may use the generic plot function, which internally selects the plot.demonoid function, which differs from plot.laplace for objects of class laplace.

For more information on object classes, see the class function.

S3 method: apply is() to objects of class blocks.

S3 method: apply is() to objects of class bmk.

S3 method: apply is() to objects of class demonoid.

S3 method: apply is() to objects of class demonoid.ppc.

S3 method: apply is() to objects of class demonoid.val.

S3 method: apply is() to objects of class hangartner.

S3 method: apply is() to objects of class heidelberger.

S3 method: apply is() to objects of class importance.

S3 method: apply is() to objects of class iterquad.

S3 method: apply is() to objects of class iterquad.ppc.

S3 method: apply is() to objects of class laplace.

S3 method: apply is() to objects of class laplace.ppc.

S3 method: apply is() to objects of class miss.

S3 method: apply is() to objects of class pmc.

S3 method: apply is() to objects of class pmc.ppc.

S3 method: apply is() to objects of class pmc.val.

S3 method: apply is() to objects of class posteriorchecks.

S3 method: apply is() to objects of class raftery.

S3 method: apply is() to objects of class rejection.

S3 method: apply is() to objects of class vb.

S3 method: apply is() to objects of class vb.ppc.

S3 method: apply is() to objects of class prescription.

S3 method: apply is() to objects of class arena.

S3 method: apply is() to objects of class consort.

S3 method: apply is() to objects of class crucible.

S3 method: apply is() to objects of class sde_model.

S3 method: apply is() to objects of class sde_fit.

Usage

is.ssm_model(x)

is.bayesfactor(x)

is.blocks(x)

is.bmk(x)

is.demonoid(x)

is.demonoid.ppc(x)

is.demonoid.val(x)

is.hangartner(x)

is.heidelberger(x)

is.importance(x)

is.iterquad(x)

is.iterquad.ppc(x)

is.laplace(x)

is.laplace.ppc(x)

is.miss(x)

is.pmc(x)

is.pmc.ppc(x)

is.pmc.val(x)

is.posteriorchecks(x)

is.raftery(x)

is.rejection(x)

is.vb(x)

is.vb.ppc(x)

is.prescription(x)

is.arena(x)

is.consort(x)

is.crucible(x)

is.sde_model(x)

is.sde_fit(x)

Arguments

x

An object that will be subjected to a logical test of its class.

Value

See Details.

Each function returns a logical value indicating whether or not the supplied object is of the corresponding class.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

See Details.

Details

Implementation of is.ssm_model. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.bayesfactor. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.blocks. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.bmk. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.demonoid. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.demonoid.ppc. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.demonoid.val. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.hangartner. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.heidelberger. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.importance. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.iterquad. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.iterquad.ppc. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.laplace. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.laplace.ppc. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.miss. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.pmc. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.pmc.ppc. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.pmc.val. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.posteriorchecks. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.raftery. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.rejection. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.vb. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.vb.ppc. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.prescription. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.arena. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.consort. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.crucible. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.sde_model. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of is.sde_fit. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

See also

Examples

if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.ssm_model
} # }
if (FALSE) { # \dontrun{
# Check class of a demonoid fit object
is.demonoid(fit)   # TRUE if fit is of class demonoid
is.laplace(fit)    # FALSE

# Check other classes
is.bayesfactor(bf) # TRUE if bf is of class bayesfactor
is.pmc(pmc_fit)    # TRUE if pmc_fit is of class pmc
is.vb(vb_fit)      # TRUE if vb_fit is of class vb
} # }

if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.blocks
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.bmk
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.demonoid
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.demonoid.ppc
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.demonoid.val
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.hangartner
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.heidelberger
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.importance
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.iterquad
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.iterquad.ppc
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.laplace
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.laplace.ppc
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.miss
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.pmc
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.pmc.ppc
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.pmc.val
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.posteriorchecks
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.raftery
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.rejection
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.vb
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.vb.ppc
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.prescription
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.arena
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.consort
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.crucible
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.sde_model
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving is.sde_fit
} # }