Convert lucifer fits to coda mcmc.list

Converts a demonoid, vb, smc, or pmc object to a mcmc.list for use with coda diagnostics or packages that accept coda objects.

S3 method: apply to_mcmc_list() to objects of class demonoid.

S3 method: apply to_mcmc_list() to objects of class vb.

S3 method: apply to_mcmc_list() to objects of class smc.

S3 method: apply to_mcmc_list() to objects of class pmc.

Usage

to_mcmc_list(x, ...)

# S3 method for class 'demonoid'
to_mcmc_list(x, use_chains = TRUE, ...)

# S3 method for class 'vb'
to_mcmc_list(x, ...)

# S3 method for class 'smc'
to_mcmc_list(x, ...)

# S3 method for class 'pmc'
to_mcmc_list(x, ...)

Arguments

x

A fitted lucifer object.

...

Additional arguments (currently unused).

use_chains

Logical. If TRUE (default) and the fit contains multiple chains, each chain is exported as a separate mcmc element. If FALSE, all samples are combined into a single chain.

Value

An object of class mcmc.list.

See Details.

See Details.

See Details.

See Details.

Details

Implementation of to_mcmc_list. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of to_mcmc_list.demonoid. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of to_mcmc_list.vb. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of to_mcmc_list.smc. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

Implementation of to_mcmc_list.pmc. Refer to the package vignettes and the cited references for a complete algorithmic and mathematical description.

See also

Examples

if (FALSE) { # \dontrun{
fit <- lucifer(Model, Data, IV, Algorithm = "NUTS")
ml <- to_mcmc_list(fit)
coda::gelman.diag(ml)
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving to_mcmc_list.demonoid
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving to_mcmc_list.vb
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving to_mcmc_list.smc
} # }
if (FALSE) { # \dontrun{
## see package vignettes for full examples involving to_mcmc_list.pmc
} # }